Move Plus-WRIO-App folder to widgets/Plus
