+++ WRIO-InternetOS/js/components/CreateDomRight.js
@@ -2,10 +2,10 @@ var React = require('react'),
     ReactDOM = require('react-dom'),
     UrlMixin = require('../mixins/UrlMixin'),
     classNames = require('classnames'),
-    ActionMenu = require('../../../Plus-WRIO-App/actions/menu'),
+    ActionMenu = require('../../../widgets/Plus/actions/menu'),
     CreateInfoTicket = require('./CreateInfoTicket'),
     CreateControlButtons = require('./CreateControlButtons'),
-    StoreMenu = require('../../../Plus-WRIO-App/stores/menu'),
+    StoreMenu = require('../../../widgets/Plus/stores/menu'),
     Reflux = require('reflux'),
     _ = require('lodash'),
     center = require('../actions/center');
