+++ widgets/Plus/stores/CrossStorageFactory.js
